Refactor cluster agent tests for maintainability
